backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

langchain
langchain-community
langchain-openai
langgraph

python-dotenv
typing-extensions

streamlit>=1.38.0
requests>=2.32.0
reportlab
openpyxl

faiss-cpu
pypdf
pandas
pydantic
pydantic-settings
python-multipart